+++ packages/components/data-table-manager/src/data-table-settings/column-manager/constants.js
@@ -1,0 +1,2 @@
+export const HIDDEN_COLUMNS_PANEL = 'hidden-columns-panel';
+export const SELECTED_COLUMNS_PANEL = 'selected-columns-panel';