+++ packages/components/data-table-manager/src/data-table-settings/density-manager/index.js
@@ -1,0 +1,1 @@
+export { default } from './density-manager';