+++ packages/components/buttons/secondary-icon-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/secondary-icon-button
 
+## 10.27.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/spacings@10.18.6
+
 ## 10.27.0
 
 ### Patch Changes